Move cov to math package
