+++ utils/math/cov.go
@@ -1,8 +1,8 @@
-package utils
+package math
 
 import "math"
 
-func COV(new float64, existingData float64, threshold float64) (bool, float64) {
+func Cov(new float64, existingData float64, threshold float64) (bool, float64) {
 	if math.Abs(new-existingData) >= threshold {
 		return true, new
 	} else {
