+++ utils/math/cov_test.go
@@ -1,4 +1,4 @@
-package utils
+package math
 
 import (
 	"fmt"
@@ -6,12 +6,12 @@ import (
 )
 
 func TestSum(t *testing.T) {
-	event, _ := COV(2, 0, 1)
+	event, _ := Cov(2, 0, 1)
 	fmt.Println(event)
-	event, _ = COV(0, 0, 1)
+	event, _ = Cov(0, 0, 1)
 	fmt.Println(event)
-	event, _ = COV(10, 9.9, 1)
+	event, _ = Cov(10, 9.9, 1)
 	fmt.Println(event)
-	event, _ = COV(-2, -1, 1)
+	event, _ = Cov(-2, -1, 1)
 	fmt.Println(event)
 }
